+++ Modules/Intent.Modules.Application.MediatR.FluentValidation/release-notes.md
@@ -1,3 +1,7 @@
+### Version 4.4.3
+
+- Fixed: Migration improvement using Roslyn instead of Regex to avoid updating incorrect IntentManaged attributes.
+
 ### Version 4.4.2
 
 - Improvement: Added support for the [new Email Address](https://github.com/IntentArchitect/Intent.Modules.NET/blob/development/Modules/Intent.Modules.Application.FluentValidation/release-notes.md#version-383) option on the Validation stereotype.